Request gates

+++ app/Http/Controllers/RoleController.php
@@ -7,17 +7,22 @@ use Symfony\Component\HttpFoundation\Response;
 use App\Models\Role;
 use App\Http\Resources\RoleResource;
 use DB;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
    
     public function index()
     {
+        Gate::authorize('view', 'roles');
+
         return RoleResource::collection(Role::all());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
         if($permissions = $request->input('permissions'))
@@ -36,11 +41,15 @@ class RoleController extends Controller
  
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
+
         return new RoleResource(Role::find($id));
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
@@ -63,6 +72,8 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
+
         DB::table('role_permission')->where('role_id', $id)->delete();
 
         Role::destroy($id);
